Refactor SwitchUserViewModel LiveData
